Split main menu into components, add eventbus and use it for refresh
