+++ frontend/src/main.ts
@@ -38,7 +38,6 @@ import App from "./App.vue"
 import router from "./router"
 
 if (location.search.startsWith("?code=")) {
-    console.log("oh look we gonna oauth")
     sessionStorage.authorizationCode = location.search.substring(6)
     location.href = "/"
 } else {
@@ -48,3 +47,29 @@ if (location.search.startsWith("?code=")) {
     app.use(router)
     app.mount("#app")
 }
+
+
+// global exports for fiddling in console
+
+import { useWinBox } from "vue-winbox"
+import { useWindowsStore } from "@/stores/windows"
+import { useWxStore } from "@/stores/wx"
+import { useNotamStore } from "@/stores/notam"
+import { usePresetStore } from "@/stores/preset"
+import { useVatsimStore } from "@/stores/vatsim"
+import { useAuthStore } from "@/stores/auth"
+import moment from "moment"
+import axios from "axios"
+
+const global = window as any
+
+global.winbox = useWinBox()
+global.windows = useWindowsStore()
+global.wx = useWxStore()
+global.notam = useNotamStore()
+global.preset = usePresetStore()
+global.vatsim = useVatsimStore()
+global.auth = useAuthStore()
+global.moment = moment
+global.axios = axios
+
